Create float conversions under Infinitable enum, deprecate top-level
The names of the top-level from_f32 and from_f64 functions are generic,
which can cause problems with the usual pattern of "use infinitable::*"

Create new functions under the corresponding Infinitable<f...> types
(using a macro to avoid duplication) and deprecate the top-level
functions.

Note that there is no type inference in this implementation.
An alternative may be to create a "FromFloat" trait to be implemented
by the Infinitable<f...> types.